App/requirements.txt
mysql-connector-python==8.0.33
Pillow==9.5.0  # For handling images in tkinter
pytest==7.3.1   # For testing